requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2022.12.7
chardet==4.0.0
Django==3.2.18
django-admin-honeypot==1.1.0
django-filter==21.1
djangorestframework==3.12.4
idna==2.10
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.6
pytest==6.2.4
pytz==2021.3
requests==2.25.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
